requirements.txt
tensorflow
pandas
scikit-learn
matplotlib
seaborn
streamlit
opencv-python-headless
Pillow